moved ddp and non-ddp tests into separate files
